+++ pennylane/gradients/gradient_transform.py
@@ -310,6 +310,162 @@ def _no_trainable_grad_legacy(tape):
     return [], lambda _: np.zeros((tape.output_dim, 0))
 
 
+def _swap_first_two_axes(grads, first_axis_size, second_axis_size):
+    """Transpose the first two axes of an iterable of iterables, returning
+    a tuple of tuples."""
+    if first_axis_size == 1:
+        return tuple(grads[0][i] for i in range(second_axis_size))
+    return tuple(
+        tuple(grads[j][i] for j in range(first_axis_size)) for i in range(second_axis_size)
+    )
+
+
+def _reorder_grad_axes_single_measure_shot_vector(grads, num_params, len_shot_vec):
+    """Reorder the axes for gradient results obtained for a tape with a single measurement from
+    a device that defined a shot vector.
+
+    The order of axes of the gradient output matches the structure outputted by jax.jacobian
+    for a tuple-valued function. Internally, this may not be the case when computing the
+    gradients, so the axes are reordered here.
+
+    The first axis always corresponds to the number of trainable parameters because the
+    parameter-shift transform defines multiple tapes each of which corresponds to a trainable
+    parameter. Those tapes are then executed using a device, which at the moment outputs results
+    with the first axis corresponding to each tape output.
+
+    The final order of axes of gradient results should be:
+    1. Shot vector
+    2. Measurements
+    3. Number of trainable parameters (Num params)
+    4. Broadcasting dimension
+    5. Measurement shape
+
+    According to the order above, the following reordering is done:
+
+    Shot vectors:
+
+        Go from
+        1. Num params
+        2. Shot vector
+        3. Measurement shape
+
+        To
+        1. Shot vector
+        2. Num params
+        3. Measurement shape
+    """
+    return _swap_first_two_axes(grads, num_params, len_shot_vec)
+
+
+def _reorder_grad_axes_multi_measure(
+    grads, num_params, num_measurements, len_shot_vec, shot_vector_multi_measure
+):
+    """Reorder the axes for gradient results obtained for a tape with multiple measurements.
+
+    The order of axes of the gradient output matches the structure outputted by jax.jacobian for
+    a tuple-valued function. Internally, this may not be the case when computing the gradients,
+    so the axes are reordered here.
+
+    The first axis always corresponds to the number of trainable parameters because the
+    parameter-shift transform defines multiple tapes each of which corresponds to a trainable
+    parameter. Those tapes are then executed using a device, which at the moment outputs
+    results with the first axis corresponding to each tape output.
+
+    The final order of axes of gradient results should be:
+    1. Shot vector
+    2. Measurements
+    3. Number of trainable parameters (Num params)
+    4. Broadcasting dimension
+    5. Measurement shape
+
+    Parameter broadcasting doesn't yet support multiple measurements, hence such cases are not
+    dealt with at the moment by this function.
+
+    According to the order above, the following reorderings are done:
+
+    A) Analytic (``shots=None``) or finite shots:
+
+        Go from
+        1. Num params
+        2. Measurements
+        3. Measurement shape
+
+        To
+        1. Measurements
+        2. Num params
+        3. Measurement shape
+
+    B) Shot vectors:
+
+        Go from
+        1. Num params
+        2. Shot vector
+        3. Measurements
+        4. Measurement shape
+
+        To
+        1. Shot vector
+        2. Measurements
+        3. Num params
+        4. Measurement shape
+    """
+    multi_param = num_params > 1
+    if not shot_vector_multi_measure:
+        new_grad = _swap_first_two_axes(grads, num_params, num_measurements)
+    else:
+        new_grad = []
+        for i in range(len_shot_vec):
+            shot_vec_grad = []
+            for j in range(num_measurements):
+                measurement_grad = []
+                for k in range(num_params):
+                    measurement_grad.append(grads[k][i][j])
+
+                measurement_grad = tuple(measurement_grad) if multi_param else measurement_grad[0]
+                shot_vec_grad.append(measurement_grad)
+            new_grad.append(tuple(shot_vec_grad))
+
+    return new_grad
+
+
+def reorder_grads(grads, tape_specs):
+    """Reorder the axes of tape gradients according to the original tape specifications.
+
+    Args:
+        grads (list[tensorlike] or list[tuple[tensorlike]] or list[tuple[tuple[tensorlike]]]:
+            Gradient entries with leading parameter axis to be reordered.
+        tape_specs (tuple): Information about the differentiated original tape in the order
+            ``(bool: single_measure, int: num_params, int: num_measurements, bool: shot_vector,
+            mixed: shots)``.
+
+    Returns:
+        tensor_like or tuple[tensor_like] or tuple[tuple[tensor_like]]: The reordered gradient
+            entries. Consider the details below for the ordering of the axes.
+
+    - If the original tape returned a single measurement and has a single parameter, the
+      first entry is returned.
+    - If the original tape returned a single measurement and a shot vector was used,
+      reorder the gradient axes with ``_reorder_grad_axes_single_measure_shot_vector``
+    - If the original tape returned multiple measurements,
+      reorder the gradient axes with ``_reorder_grad_axes_multi_measure``.
+    """
+    single_measure, num_params, num_measurements, shot_vector, shots = tape_specs
+    if single_measure and num_params == 1:
+        return grads[0]
+    len_shot_vec = _get_num_copies(shots) if shot_vector else None
+    if single_measure and shot_vector:
+        return _reorder_grad_axes_single_measure_shot_vector(grads, num_params, len_shot_vec)
+    if not single_measure:
+        grads = _reorder_grad_axes_multi_measure(
+            grads,
+            num_params,
+            num_measurements,
+            len_shot_vec,
+            shot_vector,
+        )
+    return tuple(grads)
+
+
 class gradient_transform(qml.batch_transform):
     """Decorator for defining quantum gradient transforms.
